### 2025-02-24 #7806 @kmkuesters
@csoneson Thanks for the update! I've added both a CONTRIBUTING.md and a CODE_OF_CONDUCT.md file per the suggestion. I don't have any reviewer recommendations at this time, but I'll let you know if anything comes to mind. Looking forward to the next steps!
