### 2025-02-26 #7853 @editorialbot
[REVIEW]: pooledpeaks: Open-Source Fragment Scoring and Analysis Workflow for Population Genetic Analyses of Microsatellite Markers in Pooled Samples 
**Submitting author:** @kmkuesters (<a href="http://orcid.org/0000-0003-0238-7889">Kathleen Kuesters</a>)
**Repository:** https://github.com/kmkuesters/pooledpeaks
**Branch with paper.md** (empty if default branch): master
**Version:** v1.2.2
