### 2025-02-28 #7853 @nilabhrardas
Hi @kmkuesters,

I had a chance to review the package and have a few initial comments: the paper is quite well written and easy to follow - good job! However, the README/manual included in the package only has a guide for the installation process. In the **Features** section, there is mention of the different functionalities but no reference to the exact functions. I would have liked to see the actual names of the functions listed under the individual features.

Most importantly, the README lacks any sort of actual introductory manual or guidelines for executing the functionalities of the package. I would suggest including a simulated dataset or a publicly available dataset in the package (or otherwise, providing R code for local generation by the user). Exemplar use cases of the different/novel data cleaning/manipulation/analytical functions in the package should be introduced using this dataset. Additionally, I would suggest providing results from the analysis so that the user may verify their analysis.

Lastly, I would also suggest taking advantage of the markdown features and using ` ... ` for code snippets in the installation section and the suggested examples.

In its current state, I am unable to recommend this paper for acceptance without the suggested amendments or equivalent revisions.

Best wishes.
