### 2025-03-03 #7853 @ns-rse
Hi @kmkuesters,

First pass of my review and I agree with @nilabhrardas this is a well written and easy to follow paper. I've made some suggestions for improving the `README.md` and `CONTRIBUTING.md` in a small pull request.

I think the Vignette would benefit from the example data being included in the package and being loadable and used in the walkthrough along with showing the results so users can verify they have replicate the workflow.

Perhaps a stretch goal but using the [pkgdown](https://pkgdown.r-lib.org/) library to build a website based on the Vignette would round the documentation out nicely once these mostly small issues have been tidied up.

Hope the comments are useful,

Regards,

@ns-rse
