### 2025-03-03 #7853 @kmkuesters
Dear @nilabhrardas and @ns-rse,

Thank you both for your detailed reviews and constructive feedback. I appreciate your insights and suggestions for improving the package's documentation and usability.

I have reviewed the pull request from @ns-rse and will incorporate those suggestions along with the following updates:

1. README & Documentation Improvements:
- Expanding the Features section to list function names alongside their descriptions explicitly,
- Enhancing formatting with markdown syntax for better readability.

2. Example Data & Vignette Enhancements:
- Including a simulated or publicly available dataset within the package for reproducibility.
- Updating the Vignette to use this dataset in the walkthrough, including expected results for verification.

3. Documentation Accessibility:
- Implementing updates from the PR to improve README.md and CONTRIBUTING.md.
- Exploring using pkgdown to create a documentation website as a future enhancement.

I aim to have these changes implemented by Friday, March 14th. Please let me know if you have any additional recommendations, and thanks again for your valuable input!

Best regards,
@kmkuesters
