### 2025-03-17 #7853 @ns-rse
Hi @kmkuesters

Thanks for going through things.

I figured the Vignette link in the `README.md` would benefit from pointing to the rendered version that is hosted on CRAN so have made another small PR suggesting that (and turned the DOI's for the citations into clickable links too, my editor also removed double line breaks as it links Markdown automatically so it looks like there are more changes but they are mostly cosmetic).

Whilst browsing through the recent commits I noticed you have a merge conflict in your [pooledprojects.Rproj](https://github.com/kmkuesters/pooledpeaks/blob/e5500aaf35e5706b173fc0a622a6d60e4da70b96/pooledpeaks.Rproj#L2) that will no doubt interfere when loading in RStudio so this should I think be corrected (does an R package source need a `.Rproj` file or could it be excluded by referencing in `.gitignore` and kept locally in the development environment?).

The sample data is indeed in `inst/extdata` (and has been for a while) but the Vignette doesn't make it explicitly clear that this is where it is loaded from, users unfamiliar with R may not know that `system.file("extdata", package="pooledpeaks")` is loading data that is included in the package so it would be useful to explicitly state that.

I've run through the Vignette and found the following...

0. **General Setup** `library(pooledpeaks)` is missing from the initial loading of libraries!
1. **General Setup** As someone unfamiliar with the analyses and science here it is a bit of a mystery where the `eggcount` / `Shae10` / `mic_SMMS2` / `GS600LIZ` datasets that are created under this section derive from or what they mean.
2. `fsa_batch_imp(file_path, ...)` on calling this a graphics device opens but no content is shown, is this expected behaviour? Should it be empty? If not required can it be supressed?
3. **Scoring Peaks by Marker** I get an error if, as advised, I set the `plotting = TRUE` option for either of these commands...

```r
> scores_SMMS2 <- score_markers_rev3(my.inds = fsa_data,
                                   channel = 1,
                                   channel.ladder = 5,
                                   panel = "mic_SMMS2",
                                   ladder = GS600LIZ,
                                   init.thresh = 100,
                                   ploidy = length(mic_SMMS2),
                                   shift = 1,
                                   windowL = 1,
                                   windowR= 1,
                                   left.cond = c(0, 2.5),
                                   right.cond = 0,
                                   pref = 1,
                                   plotting = TRUE
                                   )
+ + + + + + + + + + + + + + You are writing plots for all samples to the directory 'plots_scoring'.
  For faster calculation but no plots, set 'plotting = FALSE'

Scoring  mic_SMMS2 peaks
Please be patient!

Writing plots to directory 'plots_scoring'

Saved 1/12 plots
Saved 2/12 plots
Saved 3/12 plots
Saved 4/12 plots
Saved 5/12 plots
Saved 6/12 plots
Saved 7/12 plots
Saved 8/12 plots
Saved 9/12 plots
Saved 10/12 plots
Saved 11/12 plots
Saved 12/12 plots
Cleaning up...
Error in normalizePath(path, mustWork = TRUE) :
  path[1]="/home/neil/work/git/hub/ns-rse/pooledpeaks/plots_scoringsamp23.2a_I_A01_2012-07-18.fsa_ch1_rplot.pdf": No such file or directory
```

...this is because there is a `/` (directory delimiter) missing in `path[1]`, the files exist within the `plots_scoring/` directory on my system.

5. **Data Manipulation** `head(SMMS2[,1:9])` has different column headers and data compared to the example in the Vignette

```r
> head(SMMS2_IDM[, 1:9])
     X X23.2a_I_A01_2012.07.18.fsa.1 X23.2a_I_A01_2012.07.18.fsa.2 X23.2b_I_A07_2012.07.18.fsa.1 X30.3a_I_B01_2012.07.18.fsa.1
1 211                             0                             0                             0                             0
2 215                           916                             0                           657                          4273
3 219                             0                             0                             0                             0
4 231                             0                             0                             0                             0
5 235                             0                           547                             0                             0
6 239                             0                             0                             0                             0
  X30.3a_I_B01_2012.07.18.fsa.2 X30.3b_I_B07_2012.07.18.fsa.1 X30.3b_I_B07_2012.07.18.fsa.2 X33.1a_I_C01_2012.07.18.fsa.2
1                             0                             0                             0                             0
2                             0                          1443                             0                         12424
3                             0                             0                             0                             0
4                             0                             0                             0                             0
5                           419                             0                           259                             0
6                             0                             0                             0                             0
```

7. **Initial Data Manipulation** as a consequence  `head(SMMS2_IDM[,1:9])` has different column headers and data compared to the example

```r
> head(SMMS2_IDM[, 1:9])
     X X23.2a_I_A01_2012.07.18.fsa.1 X23.2a_I_A01_2012.07.18.fsa.2 X23.2b_I_A07_2012.07.18.fsa.1 X30.3a_I_B01_2012.07.18.fsa.1
1 211                             0                             0                             0                             0
2 215                           916                             0                           657                          4273
3 219                             0                             0                             0                             0
4 231                             0                             0                             0                             0
5 235                             0                           547                             0                             0
6 239                             0                             0                             0                             0
  X30.3a_I_B01_2012.07.18.fsa.2 X30.3b_I_B07_2012.07.18.fsa.1 X30.3b_I_B07_2012.07.18.fsa.2 X33.1a_I_C01_2012.07.18.fsa.2
1                             0                             0                             0                             0
2                             0                          1443                             0                         12424
3                             0                             0                             0                             0
4                             0                             0                             0                             0
5                           419                             0                           259                             0
6                             0                             0                             0                             0
```

8. **Replicate Check** Ditto underlying data is different so output of this step of processing and the resulting data frame differ.
9. **Binding Markers and Exporting Allele Frequencies** This step combines three sets of data files and writes to `.txt` but up to this point only one (`SMMS2_PCM`) has been created. As a consequence users following along will find this step fails if they explicitly follow the instructions. I also found that by default `rbind.fill()` was not found/loaded. This is part of the [`plyr`](https://rdrr.io/cran/plyr/man/rbind.fill.html) package which users won't have loaded at the start. The `plyr` package is a bit dated, although not obsolete, and  most functionality is now available in [`dplyr::bind_rows()`](https://dplyr.tidyverse.org/reference/bind.html).
10. **Calculating Gene Identity and Genetic Distance Matrices** possibly linked to differences in data but I get matrices of `NaN` for all these steps.
11. **PCA Plot** as the matrix has `NaN`/`NA` this doesn't plot
12. **Phylogenetic Tree** ditto, can't have matrix with `NaN` / `NA` so these steps don't run.

I think Vignettes and their walkthroughs should be replicable by users but currently that isn't the case.
