### 2025-03-17 #7853 @nilabhrardas
Hi @kmkuesters,

Thank you for making the suggested changes.

I encountered the same errors that @ns-rse has kindly pointed out in detail. In addition to the results from the example pipeline not being replicable, I would suggest providing information regarding the data in the README and in detail (explicitly mention if simulated/dummy data or published/publicly available data - in which case, inclusion of a link to the source) in the vignette.

I agree with @ns-rse regarding providing a link to the rendered version of the vignette in the README over the R markdown version.

Best,
@nilabhrardas
