### 2025-03-21 #7853 @kmkuesters
Dear @nilabhrardas and @ns-rse,

Thank you both again for your incredibly detailed and thoughtful feedback. I really appreciate the time and care you've taken to test the package, walk through the vignette, and suggest meaningful improvements—it’s been incredibly helpful.

I want to acknowledge the issues you've flagged, particularly around reproducibility, unclear dataset origins, and the plotting and path-related bugs. These are all valid concerns and will certainly need to be addressed.

That said, I’m in the final stretch of finishing my dissertation this month, and I’m not sure I’ll be able to carve out time to make these revisions right away. I will try to address some minor edits if I can, but realistically, I likely won’t be able to return to the package in full until after April 11th.

Thank you again for your support and understanding. I’ll follow up once I’ve had a chance to consider your suggestions.

Best regards,
@kmkuesters
