+++ vignettes/scores_Shae10_lfex.txt
@@ -1,0 +1,158 @@
+	pos	hei	wei	ID	filename
+23.2a_I_A01_2012-07-18.fsa.1	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.1
+23.2a_I_A01_2012-07-18.fsa.2	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.2
+23.2a_I_A01_2012-07-18.fsa.3	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.3
+23.2a_I_A01_2012-07-18.fsa.4	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.4
+23.2a_I_A01_2012-07-18.fsa.5	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.5
+23.2a_I_A01_2012-07-18.fsa.6	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.6
+23.2a_I_A01_2012-07-18.fsa.7	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.7
+23.2a_I_A01_2012-07-18.fsa.8	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.8
+23.2a_I_A01_2012-07-18.fsa.9	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.9
+23.2a_I_A01_2012-07-18.fsa.10	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.10
+23.2a_I_A01_2012-07-18.fsa.11	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.11
+23.2a_I_A01_2012-07-18.fsa.12	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.12
+23.2a_I_A01_2012-07-18.fsa.13	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.13
+23.2a_I_A01_2012-07-18.fsa.14	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.14
+23.2a_I_A01_2012-07-18.fsa.15	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.15
+23.2a_I_A01_2012-07-18.fsa.16	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.16
+23.2a_I_A01_2012-07-18.fsa.17	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.17
+23.2a_I_A01_2012-07-18.fsa.18	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.18
+23.2a_I_A01_2012-07-18.fsa.19	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.19
+23.2a_I_A01_2012-07-18.fsa.20	3125	915.812921225418	215	23.2a	23.2a_I_A01_2012-07-18.fsa.20
+23.2b_I_A07_2012-07-18.fsa.1	3074	657.380615485756	215	23.2b	23.2b_I_A07_2012-07-18.fsa.1
+23.2b_I_A07_2012-07-18.fsa.2	3074	657.380615485756	215	23.2b	23.2b_I_A07_2012-07-18.fsa.2
[…]
