+++ vignettes/scores_Shae10_tdfex.txt
@@ -1,0 +1,16 @@
+	23.2a	23.2b	30.3a	30.3b	33.1a	33.1b	Multiplex_set_I_Shaem.1a	Multiplex_set_I_Shaem.1b	Multiplex_set_I_Shaem.3a	Multiplex_set_I_Shaem.3b	Multiplex_set_I_Shaem.4a	Multiplex_set_I_Shaem.4b
+173	0	0	0	0	0	0	0	0	0	0	141	116
+176	0	0	0	0	0	0	0	0	0	0	171	131
+179	0	0	0	0	0	0	0	0	0	115	158	139
+182	0	0	0	0	0	0	2897	4326	4232	6122	6192	5008
+185	0	0	0	0	0	0	4876	6587	11638	17054	23975	20009
+188	0	0	0	0	0	0	2839	4088	12267	18243	31101	25679
+191	0	0	0	0	0	0	6574	9802	8511	12657	17247	12899
+194	0	0	0	0	0	0	4465	7223	4801	7868	3608	3485
+197	0	0	0	0	0	0	356	541	0	107	352	393
+200	0	0	0	0	0	0	0	0	0	0	153	131
+203	0	0	0	0	0	0	326	307	0	137	1123	772
+206	0	0	0	0	0	0	0	0	0	0	227	184
+212	0	0	0	0	119	0	0	158	0	0	0	0
+215	915.8129	657.3806	4272.9744	1443.4136	12424	3604.7478	0	0	0	0	0	0
+218	0	0	0	0	102	0	0	0	0	0	0	0
