+++ vignettes/scores_SMMS2_tdfex.txt
@@ -1,9 +1,7 @@
-	1.2a_IV_A01_2012-05-07.fsa.1	1.2a_IV_A01_2012-05-07.fsa.2	1.2a_IV_A01_2012-05-07.fsa.3	1.2a_IV_A01_2012-05-07.fsa.4	1.2a_IV_A01_2012-05-07.fsa.5	1.2a_IV_A01_2012-05-07.fsa.6	1.2a_IV_A01_2012-05-07.fsa.7	1.2a_IV_A01_2012-05-07.fsa.8	1.2a_V_A01_2012-05-07.fsa.1	1.2a_V_A01_2012-05-07.fsa.2	1.2a_V_A01_2012-05-07.fsa.3	1.2b_IV_A07_2012-05-07.fsa.1	1.2b_IV_A07_2012-05-07.fsa.2	1.2b_IV_A07_2012-05-07.fsa.3	1.2b_IV_A07_2012-05-07.fsa.4	1.2b_IV_A07_2012-05-07.fsa.5	1.2b_IV_A07_2012-05-07.fsa.6	1.2b_IV_A07_2012-05-07.fsa.7	1.2b_IV_A07_2012-05-07.fsa.8	1.2b_V_A07_2012-05-07.fsa.1	1.2b_V_A07_2012-05-07.fsa.2	1.2b_V_A07_2012-05-07.fsa.3	1.2b_V_A07_2012-05-07.fsa.4	108.3a_IV_H01_2012-05-07.fsa.1	108.3a_IV_H01_2012-05-07.fsa.2	108.3a_IV_H01_2012-05-07.fsa.3	108.3a_IV_H01_2012-05-07.fsa.4	108.3a_IV_H01_2012-05-07.fsa.5	108.3a_IV_H01_2012-05-07.fsa.6	108.3a_IV_H01_2012-05-07.fsa.7	108.3a_IV_H01_2012-05-07.fsa.8	108.3a_V_H01_2012-05-07.fsa.1	108.3a_V_H01_2012-05-07.fsa.2	108.3a_V_H01_2012-05-07.fsa.3	108.3a_V_H01_2012-05-07.fsa.4	108.3a_V_H01_2012-05-07.fsa.5	108.3b_IV_H07_2012-05-07.fsa.1	108.3b_IV_H07_2012-05-07.fsa.2	108.3b_IV_H07_2012-05-07.fsa.3	108.3b_IV_H07_2012-05-07.fsa.4	108.3b_IV_H07_2012-05-07.fsa.5	108.3b_IV_H07_2012-05-07.fsa.6	108.3b_IV_H07_2012-05-07.fsa.7	108.3b_IV_H07_2012-05-07.fsa.8	108.3b_V_H07_2012-05-07.fsa.1	108.3b_V_H07_2012-05-07.fsa.2	108.3b_V_H07_2012-05-07.fsa.3	108.3b_V_H07_2012-05-07.fsa.4	110.1a_IV_A02_2012-05-07.fsa.1	110.1a_IV_A02_2012-05-07.fsa.2	110.1a_IV_A02_2012-05-07.fsa.3	110.1a_IV_A02_2012-05-07.fsa.4	110.1a_IV_A02_2012-05-07.fsa.5	110.1a_IV_A02_2012-05-07.fsa.6	110.1a_IV_A02_2012-05-07.fsa.7	110.1a_IV_A02_2012-05-07.fsa.8	110.1a_V_A02_2012-05-08.fsa.1	110.1a_V_A02_2012-05-08.fsa.2	110.1a_V_A02_2012-05-08.fsa.3	110.1a_V_A02_2012-05-08.fsa.4	110.1a_V_A02_2012-05-08.fsa.5	110.1a_V_A02_2012-05-08.fsa.6
[…]
