### 2025-04-21 580f43b
Increment version number to 1.2.0
