### 2025-04-27 #7853 @nilabhrardas
Hi @kmkuesters,

Nice work on the suggested updates. It's basically there - just needs fixing a couple of minor errors I encountered while going through the example workflow:

In the **Binding Markers and Exporting Allele Frequencies** section, it is worth emphasizing that this particular step is optional and not part of the example workflow - I ran into an error, and it took me a couple of minutes to figure out why SMMS13_PCM and SMMS16_PCM were non-existent. Perhaps putting "Optional" in the section title would make it clearer to a first-time user? Anyhow, up to you to decide the best way of conveying the message.

In the **Loading the Dataset** section of **4. Genetic Analysis**, `gends <- LoadData("./combined3.txt")` returns an error since it directs R to look in the parent directory of the working directory (which defaults to the user's Documents directory). I would suggest changing it to `gends <- LoadData(paste0(file_path, "/combined3.txt"))` or something similar so that R searches for the file in the package directory.

That's it from my side. I wish you all the best.

@csoneson, I am happy to recommend this paper for acceptance once the above suggestions have been addressed.

Best,
@nilabhrardas
