### 2025-05-01 9f6f4e5
Vignette patch fix
