### 2025-05-01 59efdef
Increment version number to 1.2.1
