### 2025-05-02 #7853 @csoneson
Thank you @kmkuesters. I have looked at your recent commits and as far as I can tell the two suggestions raised by the reviewers have been addressed:

- Added OPTIONAL to the title of the 'Binding Markers and Exporting Allele Frequencies' section
- Changed the path to load data in the Loading the Dataset section

I will also take a quick look through the submission, and will get back to you shortly with the next steps.
