### 2025-05-02 #7853 @csoneson
@kmkuesters I have gone through your submission, and I have just a couple of minor comments regarding the paper:

- Could you please spell out 'USA' to 'United States of America' in the affiliations?
- Also in the affiliations, could you use 'Brazil' consistently (it's sometimes 'Brasil')?
- In a few places you refer to multiple references just after each other. You can combine them in one set of parentheses by enclosing them in the same `[]`, and separating them with `;`
- Please capitalize 'windows' and 'excel'

You can then generate a new proof with `@editorialbot generate pdf`.
