### 2025-05-02 #7853 @kmkuesters
Dear @csoneson,

Thanks so much for reviewing the submission and for the helpful comments. I’ve gone ahead and made the requested changes to the affiliations, references, and capitalization. I’ll generate a new proof shortly.

Please let me know if there’s anything else I should address!

Best,
@kmkuesters
