+++ inst/paper.md
@@ -17,18 +17,19 @@ authors:
     orcid: 0000-0002-1653-7836
     affiliation: 1
   - name: Jeffrey D. Kovach
-    affiliation: 2
+    orcid: 0000-0001-9051-7676
+    affiliation: "2,3"
   - name: Walter A. Blank
-    affiliation: 3
+    affiliation: 4
   - name: Lúcio M. Barbosa
     orcid: 0000-0001-7272-0515  
-    affiliation: "4,5"
+    affiliation: "5,6"
   - name: Luciano K. Silva
     orcid: 0000-0002-7739-4082
-    affiliation: 4
+    affiliation: 6
   - name: Mitermayer G. Reis
     orcid: 0000-0002-3051-9060
-    affiliation: "5,6,7"
+    affiliation: "6,7,8"
   - name: Ronald E. Blanton
     orcid: 0000-0001-6655-7336
     affiliation: 1
@@ -39,18 +40,21 @@ affiliations:
  - name: Case Western Reserve University School of Medicine, Cleveland, OH 44106, United States of America
    index: 2
    ror: 051fd9666
- - name: Independent Researcher, United States of America
+ - name: Department of Quantitative Health Sciences, Lerner Research Institute, Cleveland Clinic, Cleveland, OH, United States of America
    index: 3
- - name: Bahiana School of Medicine and Public Health, Av. Silveira Martins, n 3386, Salvador, Bahia, 41150-100, Brazil
+   ror: 03xjacd83
+ - name: Independent Researcher, United States of America
    index: 4
+ - name: Bahiana School of Medicine and Public Health, Av. Silveira Martins, n 3386, Salvador, Bahia, 41150-100, Brazil
+   index: 5
    ror: 0300yd604
  - name: Instituto Gonçalo Moniz, Fundação Oswaldo Cruz, Ministério da Saúde, Rua Waldemar Falcão, 121, Candeal, CEP 40296-710, Salvador, Bahia, Brazil
-   index: 5
- - name: Faculdade de Medicina, Universidade Federal da Bahia, Praça XV de novembro, s/n - Largo do Terreiro de Jesus, CEP 40026-010, Salvador, Bahia, Brazil
    index: 6
+ - name: Faculdade de Medicina, Universidade Federal da Bahia, Praça XV de novembro, s/n - Largo do Terreiro de Jesus, CEP 40026-010, Salvador, Bahia, Brazil
+   index: 7
    ror: 03k3p7647
  - name: Department of Epidemiology of Microbial Diseases, Yale School of Public Health, 60 College St, New Haven, Connecticut, 06510, United States of America
-   index: 7
+   index: 8
    ror: 03v76x132
 date: 1 February 2025
 bibliography: pooledpeaks_paper.bib
