### 2025-05-08 fae76b7
Increment version number to 1.2.2
