### 2025-05-09 #7853 @kmkuesters
Hi @csoneson,

- I apologize for the delay; I was waiting to hear back from my co-authors about their affiliations, which are now all correct and complete.
- The latest version has been released and is version 1.2.2.
- I have archived the package on Zenodo DOI: 10.5281/zenodo.15366707
- The author list matches in the archive and the paper as far as I can tell.
- And lastly, the software license listed is the same.

Thanks,
@kmkuesters
