### 2025-05-09 #7853 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
