### 2025-05-17 #7853 @Kevin-Mattheus-Moerman
@kmkuesters as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks I've processed. All looks good to proceed.
