### 2025-05-17 #7853 @Kevin-Mattheus-Moerman
@kmkuesters congratulations on this JOSS publication!!!!

@csoneson thanks for editing!

And a special thanks to the reviewers: @ns-rse, @nilabhrardas
