### 2025-05-19 #7853 @kmkuesters
Thank you all for your help and guidance throughout this process; it has been a delight to publish in JOSS!

Best, @kmkuesters
